Many more String changes, so that finally the String class can store its internal data as either utf8, 16 or 32 - this is controlled by a flag JUCE_STRING_UTF_TYPE. It's currently set to utf-8 by default.

+++ src/io/streams/juce_MemoryOutputStream.cpp
@@ -131,7 +131,7 @@ int MemoryOutputStream::writeFromInputStream (InputStream& source, int64 maxNumB
 
 const String MemoryOutputStream::toUTF8() const
 {
-    return String::fromUTF8 (static_cast <const char*> (getData()), getDataSize());
+    return String::fromUTF8 (static_cast <const char*> (getData()), (int) getDataSize());
 }
 
 const String MemoryOutputStream::toString() const
